really not sure
double + triple exponential matrixes are *almost* the same, precision issues
(inputs to the function are almost the same, like definitely a result from precision issue) but that seems to compound when calculating the actual se w/ matrix inversion.

confused with what tolerance is.

continue on to logTWLR? but SE is super off for two and three mixed